Documentation for our Error class, improved printing utilities, and tests for both of these. Selected deactivation of compiler optimizations. (#115)
The PR title says most of what this changes. 

The most important change is that we disable compiler optimizations for
Random123's Box-Muller transform and for our printing function. The
former change was needed to resolve a failing test for RandLAPACK. The
latter change is just to err on the side of caution.